Feat: Support Apple Silicon builds of Node (take 2) (#28)
This is second attempt at https://github.com/UrbanCompass/hnvm/pull/25, which was reverted by #26 and #27.

Changes vs. #25:

1. Only attempt to download `x64` and `arm64` builds of Node (not `amd64` or other values that `uname -m` might return that don't correspond to valid Node downloads)
3. Don't emit `package.json` to stdout when checking its validity

My speculation is that the CI failures we saw with the #25 build were caused by `uname -m` returning `amd64`, causing Node to fail to download. However, it's curious to me that no error was emitted in that case. When I try to force an invalid Node download URL on my machine, I get the error

```
tar: Error opening archive: Unrecognized archive format
```

So I can't say that I'm completely sure that this change won't introduce the same CI issue as before. 🤞